libc and x86 ieeefp: five undefined arithmetic operations the model checker names
The CBMC half had not been re-read in a long time.  Rebuilt over
lib/libc and lib/msun: 1,177 of 1,295 translation units modelled, and of
the functions the ledger and the goto model agree on, 770 checkable
unguarded (SCALAR and VOID, no stated precondition):

  BOUNDED 14  ERROR 60  FAILED 209  PROVED 425  TIMEOUT 62

Most of the 209 is four shapes.  96 division by zero, of which 30 are
libm's DELIBERATE IEEE idioms -- vzero / vzero, (double)-1 / zero,
(x - x) / (x - x) -- written to raise the invalid or divide-by-zero
exception and return the NaN or infinity the function owes its caller.
C11 6.5.5p5 makes division by zero undefined for every type and Annex F,
which FreeBSD's libm assumes throughout, defines exactly these; CBMC's
--div-by-zero-check does not know which document is in force.  54
dereference failures, every one a global or a callee's out-parameter
CBMC cannot see -- __getCurrentRuneLocale() returning NULL,
__libc_interposing[] entries, _elf_aux_info()'s out-parameter in
arc4random's _rs_initialize_fxrng.  16 memory-leak reports on functions
that allocate and return the pointer.

That left the arithmetic, and five of it were real.

  killpg(INT32_MIN, sig) returned kill(-pgid, sig).  pid_t is __int32_t
  on every architecture, so -pgid is undefined at the most negative
  value -- and there is no answer to give kill(2) either, since
  +2147483648 is not a pid_t.  No process group has that id, so it is
  ESRCH for the same reason pgid == 1 already is three lines above.

  nice(INT_MAX): prio is in [PRIO_MIN, PRIO_MAX] but incr is whatever
  the caller passed.  In long and saturated to the int range;
  setpriority(2) clamps anyway, so no result that did not already
  overflow changes.

  timezone(INT_MIN, dst): _tztab() negated its int zone in place, and
  timezone(3) names no domain for it.  In long, where every int has an
  exact negation.

  fpsetmask, on EVERY call: `_newcw |= (~_m << FP_MSKS_OFF) &
  FP_MSKS_FLD'.  fp_except_t is int on x86 -- a #define in
  x86_ieeefp.h -- so ~_m is negative for every mask with the top bit
  clear, which is every mask anyone passes, fpsetmask(0) included, and
  C11 6.5.7p4 makes a left shift of a negative value undefined whatever
  the distance.  Not an edge case: the ordinary path.  The complement is
  unsigned now and the & immediately below discards everything the
  change could affect.  The i386 copy of the same line went with it.

Same tree layout, same scope, same 770 pairs:

  before   BOUNDED 14  ERROR 60  FAILED 209  PROVED 425  TIMEOUT 62
  after    BOUNDED 15  ERROR 61  FAILED 205  PROVED 429  TIMEOUT 60

Six verdicts moved and every one is accounted for: fpsetmask, killpg,
nice and _tztab FAILED -> PROVED, __timezone_compat FAILED -> BOUNDED,
and arc4random TIMEOUT -> FAILED.  The last is not a regression -- it
timed out on the first run and the solver got further on the second,
reaching the _rs_initialize_fxrng dereference above.  A TIMEOUT that
becomes a verdict is the instrument finishing, not the tree changing,
and the same jitter is the whole of the ERROR and TIMEOUT plus-or-minus
one.

+++ hbsd/src/sys/amd64/include/ieeefp.h
@@ -156,12 +156,21 @@ __fpsetmask(fp_except_t _m)
 	__fnstcw(&_cw);
 	_p = (~_cw & FP_MSKS_FLD) >> FP_MSKS_OFF;
 	_newcw = _cw & ~FP_MSKS_FLD;
-	_newcw |= (~_m << FP_MSKS_OFF) & FP_MSKS_FLD;
+	/*
+	 * The complement in UNSIGNED. fp_except_t is `int' on x86 (a
+	 * #define in x86_ieeefp.h), so ~_m is negative for every mask
+	 * with the top bit clear -- which is every mask anyone passes,
+	 * fpsetmask(0) included -- and a left shift of a negative value
+	 * is undefined by C11 6.5.7p4 whatever the distance. The `&'
+	 * below discards everything the change could affect, so the
+	 * result is identical for all inputs.
+	 */
+	_newcw |= ((unsigned)~_m << FP_MSKS_OFF) & FP_MSKS_FLD;
 	__fnldcw(_cw, _newcw);
 	__stmxcsr(&_mxcsr);
 	/* XXX should we clear non-ieee SSE_DAZ_FLD and SSE_FZ_FLD ? */
 	_mxcsr &= ~SSE_MSKS_FLD;
-	_mxcsr |= (~_m << SSE_MSKS_OFF) & SSE_MSKS_FLD;
+	_mxcsr |= ((unsigned)~_m << SSE_MSKS_OFF) & SSE_MSKS_FLD;
 	__ldmxcsr(&_mxcsr);
 	return (_p);
 }
